+++ lib/pmdtester/builders/pmd_report_builder.rb
@@ -103,7 +103,8 @@ module PmdTester
                 "-R #{project.get_config_path(@pmd_branch_name)} " \
                 "-r #{project.get_pmd_report_path(@pmd_branch_name)} " \
                 "#{fail_on_violation} -t #{@threads} " \
-                "#{project.auxclasspath}"
+                "#{project.auxclasspath}" \
+                "#{pmd7? ? ' --no-progress' : ''}"
       start_time = Time.now
       if File.exist?(project.get_pmd_report_path(@pmd_branch_name))
         logger.warn "#{@pmd_branch_name}: Skipping PMD run - report " \
@@ -199,6 +200,10 @@ module PmdTester
     def should_use_long_cli_options
       logger.debug "PMD Version: #{@pmd_version}"
       Semver.compare(@pmd_version, '6.41.0') >= 0
+    end
+
+    def pmd7?
+      Semver.compare(@pmd_version, '7.0.0-SNAPSHOT') >= 0
     end
   end
 end